gdk3 test: Gdk::EventCorssing.{x,y} returns kind of `Float`

+++ gdk3/test/test-gdk-event.rb
@@ -351,13 +351,11 @@ class TestGdkEvent < Test::Unit::TestCase
     end
 
     def test_x
-      # TODO: Is Float better?
-      assert_kind_of(Integer, @crossing.x)
+      assert_kind_of(Float, @crossing.x)
     end
 
     def test_y
-      # TODO: Is Float better?
-      assert_kind_of(Integer, @crossing.y)
+      assert_kind_of(Float, @crossing.y)
     end
 
     def test_x_root
